requirements.txt
python-dotenv
python-telegram-bot
requests
kucoin-python
colorama
psutil
pandas
pandas-ta
questionary